tweaking, pushing to picnic

+++ setup.sh
@@ -3,24 +3,49 @@
 # make sure we have the submodules
 git submodule update --init --recursive
 
+platform='unknown'
+unamestr=$(uname)
+if [[ "$unamestr" == 'Linux' ]]; then
+   platform='linux'
+elif [[ "$unamestr" == 'Darwin' ]]; then
+   platform='osx'
+fi
+
+
 # vim setup
 if [ ! -d $HOME/.vim ]
 then
     mkdir $HOME/.vim
 fi
 cp .vim/* $HOME/.vim/*
+vim +PlugInstall +qall
 
-# FZF
-git clone --depth 1 https://github.com/junegunn/fzf.git $HOME/.fzf
-$HOME/.fzf/install
+if ! command -v fzf &> /dev/null
+then
+  # FZF
+  git clone --depth 1 https://github.com/junegunn/fzf.git $HOME/.fzf
+  $HOME/.fzf/install
+fi
 
 # Ripgrep
-curl -LO 'https://github.com/BurntSushi/ripgrep/releases/download/11.0.2/ripgrep-11.0.2-x86_64-unknown-linux-musl.tar.gz'
-tar xf ripgrep-11.0.2-*.tar.gz
-if [ ! -d $HOME/bin/rg ]
+if ! command -v rg &> /dev/null
 then
-    mkdir $HOME/bin/rg
+  curl -LO 'https://github.com/BurntSushi/ripgrep/releases/download/11.0.2/ripgrep-11.0.2-x86_64-unknown-linux-musl.tar.gz'
+  tar xf ripgrep-11.0.2-*.tar.gz
+  if [ ! -d $HOME/bin/rg ]
+  then
+      mkdir $HOME/bin/rg
+  fi
+  cp ripgrep-11.0.2-*/rg $HOME/bin/rg
+  rm ripgrep-11.0.2-*.tar.gz && rm -rf ripgrep-11.0.2-x86_64*
+  echo "export PATH=\$PATH:\$HOME/bin/rg" >> ~/.bashrc
 fi
-cp ripgrep-11.0.2-*/rg $HOME/bin/rg
-rm ripgrep-11.0.2-*.tar.gz && rm -rf ripgrep-11.0.2-x86_64*
-echo "export PATH=\$PATH:\$HOME/bin/rg" >> ~/.bashrc
+
+# silver searcher
+if ! command -v ag &> /dev/null
+then
+  if [[ "$platform" == "linux" ]]
+  then
+    sudo apt install silversearcher-ag
+  fi
+fi
